+++ CHANGELOG.md
@@ -5,6 +5,18 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Versioning based on [Semantic Versioning](https://semver.org/), with suffix
 based on [rockspec versioning](https://github.com/luarocks/luarocks/wiki/Rockspec-format).
 
+## [1.0.0-1] - UNRELEASED
+
+### Removed
+
+### Changed
+- No longer allow keywords as index fields (generated invalid Lua code)
+
+### Fixed
+- Fixed allowing keywords as variable names
+
+### Added
+
 ## [0.6.0-1] - June 16, 2023
 
 ### Changed